// src/quiz/types.ts
export enum AssessmentType {
  PRE_ASSESSMENT = 'PRE_ASSESSMENT',
  SKILL_CHECK = 'SKILL_CHECK',
}

export enum QuizStatus {
  IN_PROGRESS = 'IN_PROGRESS',
  COMPLETED = 'COMPLETED',
  ABANDONED = 'ABANDONED',
}

export enum QuestionType {
  MULTIPLE_CHOICE = 'multiple_choice',
  DESCRIPTIVE = 'descriptive',
  SCENARIO = 'scenario',
}